clippy: fix driver crates for bare-metal targets
Resolve all clippy lints in the driver tree under
x86_64-unknown-none / aarch64-unknown-none with -D warnings. The
arch-gated (cfg(target_arch=...)) and kernel-test-gated code here had
never been clippy-linted before, since host clippy can't compile it.

Dominant fixes: site-specific `// SAFETY:` comments on previously
undocumented unsafe blocks (MMIO reads/writes, DMA buffers, BAR/ECAM
config access, transmutes), plus mechanical lints (needless_range_loop,
unnecessary_cast, identity_op, manual_div_ceil, field_reassign_with_default,
doc list indentation, slow_vector_initialization -> alloc::vec!, etc.).

+++ drivers/nvidia/src/edid_aux.rs
@@ -69,7 +69,7 @@ pub unsafe fn read_edid_block(
 
     // 1. Segment select (only when block > 1).
     if block_number > 1 {
-        let seg = (block_number >> 1) as u8;
+        let seg = block_number >> 1;
         // SAFETY: caller's responsibility.
         let _ = unsafe {
             aux_xfer_retry(
@@ -84,7 +84,7 @@ pub unsafe fn read_edid_block(
         };
     }
     // 2. I²C write byte offset 0 (start of the block).
-    let byte_off = ((block_number & 1) as u8).wrapping_mul(128);
+    let byte_off = (block_number & 1).wrapping_mul(128);
     // SAFETY: same.
     let _ = unsafe {
         aux_xfer_retry(
